+++ pyznap/__init__.py
@@ -9,4 +9,4 @@
 """
 
 
-__version__ = '1.6.0'
+__version__ = 'v1.6.0+local-b2'