+++ pyznap/main.py
@@ -19,6 +19,7 @@ from .utils import read_config, create_config
 from .clean import clean_config
 from .take import take_config
 from .send import send_config
+from pyznap import __version__
 
 
 DIRNAME = os.path.dirname(os.path.abspath(__file__))
@@ -33,7 +34,8 @@ def _main():
         Exit code
     """
 
-    parser = ArgumentParser(prog='pyznap', description='ZFS snapshot tool written in python')
+    parser = ArgumentParser(prog='pyznap', 
+                            description='ZFS snapshot tool written in python {}'.format(__version__))
     parser.add_argument('-n', '--dry-run', action="store_true",
                         dest="dry_run", help="Dry-run, don't execute commands")
     parser.add_argument('-v', '--verbose', action="store_true",
@@ -82,6 +84,8 @@ def _main():
     parser_send.add_argument('--retry-interval', action="store", type=int,
                              dest='retry_interval', default=10,
                              help='interval in seconds between retries. default is 10')
+    
+    parser.epilog = "ZFS properties: [pyznap:exclude, pyznap:max_size]"
 
     if len(sys.argv)==1:
         parser.print_help(sys.stderr)
@@ -93,7 +97,7 @@ def _main():
                         datefmt='%b %d %H:%M:%S', stream=sys.stdout)
     logger = logging.getLogger(__name__)
 
-    logger.info('Starting pyznap...')
+    logger.info('Starting pyznap {}...'.format(__version__))
 
     if args.command in ('snap', 'send'):
         config_path = args.config if args.config else os.path.join(CONFIG_DIR, 'pyznap.conf')